requirements.txt
flask_jwt_extended
python-telegram-bot
pyyaml
bs4
requests